% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severance_palettes.R
\docType{data}
\name{severance_palettes}
\alias{severance_palettes}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 6.
}
\usage{
severance_palettes
}
\description{
Use names(severance) to return all possible palette names. Current choices are:
\code{Dinner}, \code{Hell}, \code{HideAndSeek}, \code{Jazz01}, \code{Jazz02}, \code{TheYouYouAre}
}
\keyword{datasets}
